ok , now everything works fine TODO: system objects and join qml with c++

+++ modules/backend/converter/Converter.cpp
@@ -23,7 +23,9 @@ QString UnsignedNumericValueConverter::getValue(const QByteArray& bytes) const
     uint64_t temp_result{0};
 
     for(int i{0}; i < bytes.size() ; ++i)
-        temp_result |= ( static_cast<uint8_t>(bytes[i]) << ( BITS_IN_BYTE * i ) );
+        temp_result |=   (uint64_t)( (uint8_t)bytes[i] )  << ( BITS_IN_BYTE * i ) ;
+    //                        ^           ^--- conversion to unsigned ( pure byte)
+    //                        | - extending type to 64-bit to proper bitwise shifting
 
     return QString::number(temp_result);
 }
